+++ lib/code/module/system/ui/notice/notice_add_edit_page.dart
@@ -65,27 +65,33 @@ class NoticeAddEditPage extends AppBaseStatelessWidget<_NoticeAddEditVm> {
                         ? S.current.sys_label_notice_add
                         : S.current.sys_label_notice_edit),
                     actions: [
-                      IconButton(
-                          onPressed: () {
-                            getVm().onSave();
-                          },
-                          icon: Icon(Icons.save)),
-                      if (sysNotice != null)
+                      if ((globalVm.userShareVm.hasPermiAny(["system:notice:edit"]) &&
+                              sysNotice != null) ||
+                          (globalVm.userShareVm.hasPermiAny(["system:notice:add"]) &&
+                              sysNotice == null))
+                        IconButton(
+                            onPressed: () {
+                              getVm().onSave();
+                            },
+                            icon: Icon(Icons.save)),
+                      if (sysNotice != null &&
+                          globalVm.userShareVm.hasPermiAny(["system:notice:remove"]))
                         PopupMenuButton(itemBuilder: (context) {
                           return [
-                            PopupMenuItem(
-                              child: Text(S.current.action_delete),
-                              onTap: () {
-                                FastDialogUtils.showDelConfirmDialog(context,
-                                    contentText: TextUtil.format(
-                                        S.current.sys_label_notice_del_prompt,
-                                        [sysNotice?.noticeTitle ?? ""])).then((confirm) {
-                                  if (confirm == true) {
-                                    getVm().onDelete();
-                                  }
-                                });
-                              },
-                            )
+                            if (globalVm.userShareVm.hasPermiAny(["system:notice:remove"]))
+                              PopupMenuItem(
+                                child: Text(S.current.action_delete),
+                                onTap: () {
+                                  FastDialogUtils.showDelConfirmDialog(context,
+                                      contentText: TextUtil.format(
+                                          S.current.sys_label_notice_del_prompt,
+                                          [sysNotice?.noticeTitle ?? ""])).then((confirm) {
+                                    if (confirm == true) {
+                                      getVm().onDelete();
+                                    }
+                                  });
+                                },
+                              )
                           ];
                         })
                     ],
